MinerU/requirements.txt
flask>=2.0.0
boto3>=1.26.0
pillow>=9.0.0
psutil>=5.9.0
botocore>=1.29.0
